Refactor the flower layout to move scaling flowers into outer component
This will help us be able to position other elements relative to the flowers in the future like marker number indicators and bees.
