+++ src/app/components/flower-layout-item/flower-layout-item.component.scss
@@ -1,12 +1,5 @@
-// :host {
-//   position: absolute;
-//   z-index: 1;
-
 img {
-  width: 100%;
-  //position: relative;
-  //display: inline;
   display: block;
   filter: drop-shadow(0 -5px 15px #f0f1eb);
+  width: 100%;
 }
-
