+++ src/app/components/flower-layout-item/flower-layout-item.component.ts
@@ -1,12 +1,6 @@
 import { Component, OnInit, Input, OnChanges } from '@angular/core';
 import { trigger, animate, transition, style, state} from '@angular/animations';
 
-export interface FlowerLayoutItem {
-  imgSrc: string;
-  active: boolean;
-  scale: number;
-  alt: string;
-}
 
 
 @Component({
@@ -33,7 +27,10 @@ export interface FlowerLayoutItem {
 })
 export class FlowerLayoutItemComponent implements OnInit, OnChanges {
 
-  @Input() item: FlowerLayoutItem;
+  @Input() imgSrc: string;
+  @Input() active: boolean;
+  @Input() imgAlt: string;
+
 
   constructor() { }
 
@@ -44,9 +41,5 @@ export class FlowerLayoutItemComponent implements OnInit, OnChanges {
 
   }
 
-  calculateScale(scale: number) {
-    // Normalize scale
-    return ((scale - 1) * 0.2) + 1;
-  }
 
 }
