+++ src/app/components/flower-layout/flower-layout.component.ts
@@ -1,6 +1,14 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { trackByIndex } from 'src/app/utils/array-utils';
-import { FlowerLayoutItem } from '../flower-layout-item/flower-layout-item.component';
+
+export interface FlowerLayoutItem {
+  imgSrc: string;
+  active: boolean;
+  scale: number;
+  alt: string;
+}
+
+const flowerWidth = 0.1;
 
 @Component({
   selector: 'app-flower-layout',
@@ -18,6 +26,13 @@ export class FlowerLayoutComponent implements OnInit {
   constructor() { }
 
   ngOnInit() {
+  }
+
+
+  // Base width of each flower is 10%, flowers may have their own scaling
+  calculateFlowerSize(scale: number) {
+    // Normalize scale
+    return (((scale - 1) * 0.2) + 1) * flowerWidth;
   }
 
 }
